[major] Deprecating ecs module
Currently, automatic credential refresh is handled already by ECS / boto. The intended purpose of the ECS module was for local testing or to accommodate bespoke endpoints by modulating the constants in `ecs.py`. However, it is far easier for users to instead use `custom.py`. This makes sense. Users with complicated authentication flows are more likely to find value from the custom module. This also reduces the overall complexity of boto3-refresh-session, which is attractive for users. Hence, I have decided to drop the ECS module, as it appears superfluous when the custom module does what users want far better and simpler.