[major] Added MFA support (#83)
This PR adds support for MFA.

Given the number of token providers available, as well as the fact this code is a backend dependency, priority was given to callable objects which return a token code (string). Users may, if they wish, inject TokenCode into the parameters fed to STS.Client.assume_role via RefreshableSession.assume_role_kwargs["TokenCode"], but this strategy is not recommended as it requires token refresh before temporary credential refresh, which would be tricky and fragile. In any case, users will be required to provide SerialNumber to assume_role_kwargs so long as mfa_token_provider is set or, conversely, if mfa_token_provider is not set but TokenCode is. The core motivation of this addition to boto3-refresh-session was to further accommodate common use cases among developers and teams, as well as to make MFA support easier for users (and brs developers) when the brsCLI is eventually released. Due to the fact that users are responsible for crafting their own callable object, no tests were included inboto3-refresh-session` for unit testing this new feature, as any test would merely check if the callable returned a string of length N.